Sanitize the station name in the save dialog input field
We detect characters not allowed by file system and disable Save button
in onTextChanged, but the initial input is inserted as-is, so the user
can press Save button and bypass the check. As a result, renaming the
recording file fails silently.

Also, listening the recording doesn't work because we propagate the new
recording name, that points to a file that doesn't exist because not
renamed.

Signed-off-by: iusmac <iusico.maxim@libero.it>